# src/main/java/com/probestack/forgestudio/design/api/CoverageApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.CoverageEntry;
import com.probestack.forgestudio.design.model.ErrorResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-14T07:15:01.769029535Z[GMT]")@Validated
@Tag(name = "Coverage", description = "Supported country and dataset metadata")
public interface CoverageApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /coverage : List supported countries
     * Returns the countries covered by the validation datasets and the freshness of each dataset.
     *
     * @return Coverage fetched successfully (status code 200)
     *         or Missing or invalid API Key. (status code 401)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "listCoverage",
        summary = "List supported countries",
        description = "Returns the countries covered by the validation datasets and the freshness of each dataset.",
        tags = { "Coverage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Coverage fetched successfully", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CoverageEntry.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/coverage",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<CoverageEntry>> listCoverage(
        
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"country\" : \"US\", \"datasetVersion\" : \"USPS-2026-07\", \"lastRefreshedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"countryName\" : \"United States\", \"supportsGeocoding\" : true }, { \"country\" : \"US\", \"datasetVersion\" : \"USPS-2026-07\", \"lastRefreshedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"countryName\" : \"United States\", \"supportsGeocoding\" : true } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}