# src/main/java/com/probestack/forgestudio/design/api/ValidationApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.ValidateAddressRequest;
import com.probestack.forgestudio.design.model.ValidationResult;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-14T07:15:01.769029535Z[GMT]")@Validated
@Tag(name = "Validation", description = "Validate and normalise addresses")
public interface ValidationApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /validate : Validate an address
     * Checks an address against the postal authority dataset and returns a normalised form with a confidence score.
     *
     * @param validateAddressRequest  (required)
     * @return Address validated successfully (status code 200)
     *         or Invalid request payload or parameters. (status code 400)
     *         or Missing or invalid API Key. (status code 401)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "validateAddress",
        summary = "Validate an address",
        description = "Checks an address against the postal authority dataset and returns a normalised form with a confidence score.",
        tags = { "Validation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Address validated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ValidationResult.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/validate",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ValidationResult> validateAddress(
        @Parameter(name = "ValidateAddressRequest", description = "", required = true) @Valid @RequestBody ValidateAddressRequest validateAddressRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"valid\" : true, \"normalised\" : { \"country\" : \"US\", \"city\" : \"NEW YORK\", \"postalCode\" : \"10001-1234\", \"latitude\" : 40.750742, \"state\" : \"NY\", \"line2\" : \"APT 4B\", \"line1\" : \"123 MAIN ST\", \"longitude\" : -73.993439 }, \"corrections\" : [ \"postalCode corrected from 10011 to 10001\" ], \"confidence\" : 0.97, \"deliverable\" : true }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}